+++ make.mk
@@ -108,7 +108,6 @@ SRCS += \
   $(TOP)/watch-library/hardware/watch/watch_extint.c \
   $(TOP)/watch-library/hardware/watch/watch_led.c \
   $(TOP)/watch-library/hardware/watch/watch_buzzer.c \
-  $(TOP)/watch-library/hardware/watch/watch_buzzer_seq.c \
   $(TOP)/watch-library/hardware/watch/watch_adc.c \
   $(TOP)/watch-library/hardware/watch/watch_gpio.c \
   $(TOP)/watch-library/hardware/watch/watch_i2c.c \
@@ -182,7 +181,6 @@ SRCS += \
   $(TOP)/watch-library/simulator/watch/watch_extint.c \
   $(TOP)/watch-library/simulator/watch/watch_led.c \
   $(TOP)/watch-library/simulator/watch/watch_buzzer.c \
-  $(TOP)/watch-library/simulator/watch/watch_buzzer_seq.c \
   $(TOP)/watch-library/simulator/watch/watch_adc.c \
   $(TOP)/watch-library/simulator/watch/watch_gpio.c \
   $(TOP)/watch-library/simulator/watch/watch_i2c.c \
