+++ watch-library/simulator/watch/watch_buzzer.c
@@ -30,6 +30,74 @@
 static bool buzzer_enabled = false;
 static uint32_t buzzer_period;
 
+void cb_watch_buzzer_seq(void);
+
+static uint16_t _seq_position;
+static int8_t _tone_ticks, _repeat_counter;
+static int8_t *_sequence;
+static void (*_cb_finished)(void);
+
+void watch_buzzer_play_sequence(int8_t *note_sequence, void (*callback_on_end)(void)) {
+    watch_rtc_disable_periodic_callback(32);
+    watch_set_buzzer_off();
+    _sequence = note_sequence;
+    _cb_finished = callback_on_end;
+    _seq_position = 0;
+    _tone_ticks = 0;
+    _repeat_counter = -1;
+    // prepare buzzer
+    watch_enable_buzzer();
+    // register 32 hz callback
+    watch_rtc_register_periodic_callback(cb_watch_buzzer_seq, 32);
+}
+
+void cb_watch_buzzer_seq(void) {
+    // callback for reading the note sequence
+    if (_tone_ticks == 0) {
+        if (_sequence[_seq_position] < 0 && _sequence[_seq_position + 1]) {
+            // repeat indicator
+            if (_repeat_counter == -1) {
+                // first encounter: load repeat counter
+                _repeat_counter = _sequence[_seq_position + 1];
+            } else _repeat_counter--;
+            if (_repeat_counter > 0)
+                // rewind
+                if (_seq_position > _sequence[_seq_position] * -2)
+                    _seq_position += _sequence[_seq_position] * 2;
+                else
+                    _seq_position = 0;
+            else {
+                // continue
+                _seq_position += 2;
+                _repeat_counter = -1;
+            }
+        }
+        if (_sequence[_seq_position] && _sequence[_seq_position + 1]) {
+            // read note
+            BuzzerNote note = _sequence[_seq_position];
+            if (note == BUZZER_NOTE_REST) {
+                watch_set_buzzer_off();
+            } else {
+                watch_set_buzzer_period(NotePeriods[note]);
+                watch_set_buzzer_on();
+            }
+            // set duration ticks and move to next tone
+            _tone_ticks = _sequence[_seq_position + 1];
+            _seq_position += 2;
+        } else {
+            // end the sequence
+            watch_buzzer_abort_sequence();
+            if (_cb_finished) _cb_finished();
+        }
+    } else _tone_ticks--;
+}
+
+void watch_buzzer_abort_sequence(void) {
+    // ends/aborts the sequence
+    watch_rtc_disable_periodic_callback(32);
+    watch_set_buzzer_off();
+}
+
 void watch_enable_buzzer(void) {
     buzzer_enabled = true;
     buzzer_period = NotePeriods[BUZZER_NOTE_A4];
